refactor light node code
