+++ client/src/archive/mod.rs
@@ -9,11 +9,10 @@ pub use crate::configuration::Configuration;
 use blockgen::BlockGenerator;
 
 use cfxcore::{
-    genesis, light_protocol::QueryProvider, statistics::Statistics,
-    storage::StorageManager, sync::SyncPhaseType,
-    transaction_pool::DEFAULT_MAX_BLOCK_GAS_LIMIT, vm_factory::VmFactory,
-    ConsensusGraph, SynchronizationGraph, SynchronizationService,
-    TransactionPool, WORKER_COMPUTATION_PARALLELISM,
+    genesis, statistics::Statistics, storage::StorageManager,
+    sync::SyncPhaseType, transaction_pool::DEFAULT_MAX_BLOCK_GAS_LIMIT,
+    vm_factory::VmFactory, ConsensusGraph, LightProvider, SynchronizationGraph,
+    SynchronizationService, TransactionPool, WORKER_COMPUTATION_PARALLELISM,
 };
 
 use crate::rpc::{
@@ -201,7 +200,7 @@ impl ArchiveClient {
             Arc::new(network)
         };
 
-        let light_provider = Arc::new(QueryProvider::new(
+        let light_provider = Arc::new(LightProvider::new(
             consensus.clone(),
             sync_graph.clone(),
             Arc::downgrade(&network),
