+++ core/src/light_protocol/common/ledger_proof.rs
