+++ core/src/light_protocol/provider.rs
@@ -9,13 +9,24 @@ use rlp::Rlp;
 use std::sync::{Arc, Weak};
 
 use cfx_types::H256;
-use primitives::{
-    BlockHeader, EpochNumber, Receipt, SignedTransaction, StateRoot,
-    TransactionWithSignature,
-};
+use primitives::{SignedTransaction, TransactionWithSignature};
 
 use crate::{
     consensus::ConsensusGraph,
+    light_protocol::{
+        common::{LedgerInfo, Peers, Validate},
+        handle_error,
+        message::{
+            msgid, BlockHashes as GetBlockHashesResponse,
+            BlockHeaders as GetBlockHeadersResponse, GetBlockHashesByEpoch,
+            GetBlockHeaders, GetReceipts, GetStateEntry, GetStateRoot, GetTxs,
+            NewBlockHashes, NodeType, Receipts as GetReceiptsResponse,
+            SendRawTx, StateEntry as GetStateEntryResponse,
+            StateRoot as GetStateRootResponse, StatusPing, StatusPong,
+            Txs as GetTxsResponse,
+        },
+        Error, ErrorKind, LIGHT_PROTOCOL_ID, LIGHT_PROTOCOL_VERSION,
+    },
     message::{Message, MsgId},
     network::{
         throttling::THROTTLING_SERVICE, NetworkContext, NetworkProtocolHandler,
@@ -24,31 +35,9 @@ use crate::{
     parameters::light::{
         MAX_EPOCHS_TO_SEND, MAX_HEADERS_TO_SEND, MAX_TXS_TO_SEND,
     },
-    statedb::StateDb,
-    storage::{
-        state::{State, StateTrait},
-        state_manager::StateManagerTrait,
-        SnapshotAndEpochIdRef, StateProof,
-    },
     sync::SynchronizationGraph,
     TransactionPool,
 };
-
-use super::{
-    handle_error,
-    message::{
-        msgid, BlockHashes as GetBlockHashesResponse,
-        BlockHeaders as GetBlockHeadersResponse, GetBlockHashesByEpoch,
-        GetBlockHeaders, GetReceipts, GetStateEntry, GetStateRoot, GetTxs,
-        NewBlockHashes, NodeType, Receipts as GetReceiptsResponse,
-        ReceiptsWithProof, SendRawTx, StateEntry as GetStateEntryResponse,
-        StateRoot as GetStateRootResponse, StateRootWithProof, StatusPing,
-        StatusPong, Txs as GetTxsResponse,
-    },
-    peers::Peers,
-    Error, ErrorKind, LIGHT_PROTOCOL_ID, LIGHT_PROTOCOL_VERSION,
-};
-use crate::parameters::consensus::DEFERRED_STATE_EPOCH_COUNT;
 
 #[derive(Default)]
 pub struct LightPeerState {
@@ -56,12 +45,15 @@ pub struct LightPeerState {
     pub protocol_version: u8,
 }
 
-pub struct QueryProvider {
+pub struct Provider {
     // shared consensus graph
     consensus: Arc<ConsensusGraph>,
 
     // shared synchronization graph
     graph: Arc<SynchronizationGraph>,
+
+    // helper API for retrieving ledger information
+    ledger: LedgerInfo,
 
     // shared network service
     // NOTE: use weak pointer in order to avoid circular references
@@ -72,22 +64,29 @@ pub struct QueryProvider {
 
     // shared transaction pool
     tx_pool: Arc<TransactionPool>,
+
+    // helper API for validating ledger and state information
+    validate: Validate,
 }
 
-impl QueryProvider {
+impl Provider {
     pub fn new(
         consensus: Arc<ConsensusGraph>, graph: Arc<SynchronizationGraph>,
         network: Weak<NetworkService>, tx_pool: Arc<TransactionPool>,
     ) -> Self
     {
+        let ledger = LedgerInfo::new(consensus.clone());
         let peers = Peers::new();
+        let validate = Validate::new(consensus.clone());
 
-        QueryProvider {
+        Provider {
             consensus,
             graph,
+            ledger,
             network,
             peers,
             tx_pool,
+            validate,
         }
     }
 
@@ -101,7 +100,7 @@ impl QueryProvider {
                 &[LIGHT_PROTOCOL_VERSION],
             )
             .map_err(|e| {
-                format!("failed to register protocol QueryProvider: {:?}", e)
+                format!("failed to register protocol Provider: {:?}", e)
             })
     }
 
@@ -161,161 +160,6 @@ impl QueryProvider {
     }
 
     #[inline]
-    fn pivot_hash_of(&self, epoch: u64) -> Result<H256, Error> {
-        let epoch = EpochNumber::Number(epoch);
-        Ok(self.consensus.get_hash_from_epoch_number(epoch)?)
-    }
-
-    #[inline]
-    fn pivot_header_of(&self, epoch: u64) -> Result<Arc<BlockHeader>, Error> {
-        let pivot = self.pivot_hash_of(epoch)?;
-        let header = self.consensus.data_man.block_header_by_hash(&pivot);
-        header.ok_or(ErrorKind::InternalError.into())
-    }
-
-    #[inline]
-    fn block_hashes_in(&self, epoch: u64) -> Result<Vec<H256>, Error> {
-        let epoch = EpochNumber::Number(epoch);
-        Ok(self.consensus.get_block_hashes_by_epoch(epoch)?)
-    }
-
-    #[inline]
-    fn headers_needed_to_verify(&self, epoch: u64) -> Result<Vec<u64>, Error> {
-        // find the first header that can verify the state root requested
-        let witness = self.consensus.first_epoch_with_correct_state_of(epoch);
-
-        let witness = match witness {
-            Some(epoch) => epoch,
-            None => {
-                warn!("Unable to produce state proof for epoch {}", epoch);
-                return Err(ErrorKind::UnableToProduceProof.into());
-            }
-        };
-
-        let blame = self.pivot_header_of(witness)?.blame() as u64;
-
-        // assumption: the state root requested can be verified by the witness
-        assert!(witness >= epoch + DEFERRED_STATE_EPOCH_COUNT);
-        assert!(witness <= epoch + DEFERRED_STATE_EPOCH_COUNT + blame);
-
-        // assumption: the witness header is correct
-        // i.e. it does not blame blocks at or before the genesis block
-        assert!(witness > blame);
-
-        // collect all header heights that were used to compute DSR of `witness`
-        Ok((0..(blame + 1)).map(|ii| witness - ii).collect())
-    }
-
-    #[inline]
-    fn state_of(&self, epoch: u64) -> Result<State, Error> {
-        let pivot = self.pivot_hash_of(epoch)?;
-
-        let state = self
-            .consensus
-            .data_man
-            .storage_manager
-            .get_state_no_commit(SnapshotAndEpochIdRef::new(&pivot, None));
-
-        match state {
-            Ok(Some(state)) => Ok(state),
-            _ => Err(ErrorKind::InternalError.into()),
-        }
-    }
-
-    #[inline]
-    fn state_root_of(&self, epoch: u64) -> Result<StateRoot, Error> {
-        match self.state_of(epoch)?.get_state_root() {
-            Ok(Some(root)) => Ok(root.state_root),
-            _ => Err(ErrorKind::InternalError.into()),
-        }
-    }
-
-    #[inline]
-    fn correct_deferred_state_root_hash_of(
-        &self, height: u64,
-    ) -> Result<H256, Error> {
-        let epoch = height.saturating_sub(DEFERRED_STATE_EPOCH_COUNT);
-        let root = self.state_root_of(epoch)?;
-        Ok(root.compute_state_root_hash())
-    }
-
-    #[inline]
-    fn state_root_with_proof_of(
-        &self, epoch: u64,
-    ) -> Result<StateRootWithProof, Error> {
-        let root = self.state_root_of(epoch)?;
-
-        let proof = self
-            .headers_needed_to_verify(epoch)?
-            .into_iter()
-            .map(|h| self.correct_deferred_state_root_hash_of(h))
-            .collect::<Result<Vec<H256>, Error>>()?;
-
-        Ok(StateRootWithProof { root, proof })
-    }
-
-    #[inline]
-    fn epoch_receipts_of(
-        &self, epoch: u64,
-    ) -> Result<Vec<Vec<Receipt>>, Error> {
-        let pivot = self.pivot_hash_of(epoch)?;
-        let hashes = self.block_hashes_in(epoch)?;
-
-        hashes
-            .into_iter()
-            .map(|h| {
-                self.consensus
-                    .data_man
-                    .block_results_by_hash_with_epoch(&h, &pivot, false)
-                    .map(|res| (*res.receipts).clone())
-                    .ok_or(ErrorKind::InternalError.into())
-            })
-            .collect()
-    }
-
-    #[inline]
-    fn correct_deferred_receipts_root_hash_of(
-        &self, height: u64,
-    ) -> Result<H256, Error> {
-        let epoch = height.saturating_sub(DEFERRED_STATE_EPOCH_COUNT);
-        let pivot = self.pivot_hash_of(epoch)?;
-
-        self.consensus
-            .data_man
-            .get_epoch_execution_commitments(&pivot)
-            .map(|c| c.receipts_root)
-            .ok_or(ErrorKind::InternalError.into())
-    }
-
-    #[inline]
-    fn receipts_with_proof_of(
-        &self, epoch: u64,
-    ) -> Result<ReceiptsWithProof, Error> {
-        let receipts = self.epoch_receipts_of(epoch)?;
-
-        let proof = self
-            .headers_needed_to_verify(epoch)?
-            .into_iter()
-            .map(|h| self.correct_deferred_receipts_root_hash_of(h))
-            .collect::<Result<Vec<H256>, Error>>()?;
-
-        Ok(ReceiptsWithProof { receipts, proof })
-    }
-
-    #[inline]
-    fn state_entry_at(
-        &self, epoch: u64, key: &Vec<u8>,
-    ) -> Result<(Option<Vec<u8>>, StateProof), Error> {
-        let state = self.state_of(epoch)?;
-
-        let (value, proof) = StateDb::new(state)
-            .get_raw_with_proof(key)
-            .or(Err(ErrorKind::InternalError))?;
-
-        let value = value.map(|x| x.to_vec());
-        Ok((value, proof))
-    }
-
     fn tx_by_hash(&self, hash: H256) -> Option<SignedTransaction> {
         if let Some(info) = self.consensus.get_transaction_info_by_hash(&hash) {
             return Some(info.0);
@@ -353,20 +197,6 @@ impl QueryProvider {
     }
 
     #[inline]
-    fn validate_genesis_hash(&self, genesis: H256) -> Result<(), Error> {
-        match self.consensus.data_man.true_genesis_block.hash() {
-            h if h == genesis => Ok(()),
-            h => {
-                debug!(
-                    "Genesis mismatch (ours: {:?}, theirs: {:?})",
-                    h, genesis
-                );
-                Err(ErrorKind::GenesisMismatch.into())
-            }
-        }
-    }
-
-    #[inline]
     fn validate_peer_type(&self, node_type: &NodeType) -> Result<(), Error> {
         match node_type {
             NodeType::Light => Ok(()),
@@ -381,7 +211,7 @@ impl QueryProvider {
         info!("on_status peer={:?} status={:?}", peer, status);
 
         self.validate_peer_type(&status.node_type)?;
-        self.validate_genesis_hash(status.genesis_hash)?;
+        self.validate.genesis_hash(status.genesis_hash)?;
 
         if let Err(e) = self.send_status(io, peer) {
             warn!("Failed to send status to peer={:?}: {:?}", peer, e);
@@ -402,8 +232,8 @@ impl QueryProvider {
         info!("on_get_state_root req={:?}", req);
         let request_id = req.request_id;
 
-        let pivot_hash = self.pivot_hash_of(req.epoch)?;
-        let state_root = self.state_root_with_proof_of(req.epoch)?;
+        let pivot_hash = self.ledger.pivot_hash_of(req.epoch)?;
+        let state_root = self.ledger.state_root_with_proof_of(req.epoch)?;
 
         let msg: Box<dyn Message> = Box::new(GetStateRootResponse {
             request_id,
@@ -422,9 +252,9 @@ impl QueryProvider {
         info!("on_get_state_entry req={:?}", req);
         let request_id = req.request_id;
 
-        let pivot_hash = self.pivot_hash_of(req.epoch)?;
-        let state_root = self.state_root_with_proof_of(req.epoch)?;
-        let (entry, proof) = self.state_entry_at(req.epoch, &req.key)?;
+        let pivot_hash = self.ledger.pivot_hash_of(req.epoch)?;
+        let state_root = self.ledger.state_root_with_proof_of(req.epoch)?;
+        let (entry, proof) = self.ledger.state_entry_at(req.epoch, &req.key)?;
         let entry = entry.map(|x| x.to_vec());
 
         let msg: Box<dyn Message> = Box::new(GetStateEntryResponse {
@@ -528,8 +358,8 @@ impl QueryProvider {
         info!("on_get_receipts req={:?}", req);
         let request_id = req.request_id;
 
-        let pivot_hash = self.pivot_hash_of(req.epoch)?;
-        let receipts = self.receipts_with_proof_of(req.epoch)?;
+        let pivot_hash = self.ledger.pivot_hash_of(req.epoch)?;
+        let receipts = self.ledger.receipts_with_proof_of(req.epoch)?;
 
         let msg: Box<dyn Message> = Box::new(GetReceiptsResponse {
             request_id,
@@ -619,7 +449,7 @@ impl QueryProvider {
     }
 }
 
-impl NetworkProtocolHandler for QueryProvider {
+impl NetworkProtocolHandler for Provider {
     fn initialize(&self, _io: &dyn NetworkContext) {}
 
     fn on_message(&self, io: &dyn NetworkContext, peer: PeerId, raw: &[u8]) {
